fixed cloudinary bug

+++ src/utils/cloudinary.js
@@ -1,10 +1,14 @@
+import dotenv from "dotenv"
+dotenv.config({
+    path: './.env'
+})
 import { v2 as cloudinary} from "cloudinary";
 import fs from "fs";
 
 cloudinary.config({ 
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-  api_key: process.env.CLOUDINARY_API_KEY ? "FOUND" : "MISSING", 
-  api_secret: process.env.CLOUDINARY_API_SECRET ? "FOUND" : "MISSING"
+  api_key: process.env.CLOUDINARY_API_KEY, 
+  api_secret: process.env.CLOUDINARY_API_SECRET
 }); 
 
 const uploadOnCloudinary = async (localFilePath) => {
@@ -15,12 +19,17 @@ const uploadOnCloudinary = async (localFilePath) => {
             resource_type: "auto"
         });
         //file has been uploaded successfully
-        console.log("File is uploaded on cloudinary", response.url );
+        // console.log("File is uploaded on cloudinary", response.url );
+        fs.unlinkSync(localFilePath)
         return response;
 
     } catch(error){
+    console.log("CLOUDINARY ERROR:", error);
+
+    if (fs.existsSync(localFilePath)) {
         fs.unlinkSync(localFilePath);
-        return null;
+    }
+    return null;
 };
     
 }
